pkg/integrators/first_model_grule/geofences/requirements.txt
mysql-connector-python==9.5.0
requests==2.32.5
